% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{catch_member_info_by_alph}
\alias{catch_member_info_by_alph}
\title{Grabs member information by iterating through the alphabet.}
\usage{
catch_member_info_by_alph(parliments_source)
}
\arguments{
\item{parliments_source}{A string representing the source URL for parliament members' information.}
}
\value{
No direct return value, but calls other functions internally to fetch member information.
}
\description{
This function reads an HTML page containing member information and iterates
through each letter of the alphabet to collect member details URLs.
}
\examples{
# Example usage:
# Note: Replace with a valid URL for actual use
catch_member_info_by_alph("https://example.com/members/")

}
