% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{catch_member_info_by_party}
\alias{catch_member_info_by_party}
\title{Catch Member Information by Party}
\usage{
catch_member_info_by_party(parliments_source)
}
\arguments{
\item{parliments_source}{A string representing the source URL for parliament members' information.}
}
\value{
A tibble containing the collected member information.
}
\description{
This function collects member information by iterating through a list of political parties.
}
<<<<<<< HEAD
\examples{
# Example usage:
# Note: Replace with a valid URL for actual use
catch_member_info_by_party("https://example.com/members/")

}
=======
>>>>>>> c70c6e3b195ab3ed0ee9e1f3ccd22fdfd1a015ea
