% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{draw_member_age_hist}
\alias{draw_member_age_hist}
\title{Plot Member Age Distribution Histogram}
\usage{
draw_member_age_hist(input, output)
}
\arguments{
\item{input}{A list containing parameters. \code{party_name} is a string representing the name of the party.}

\item{output}{A Shiny output object used to render the histogram.}
}
\description{
This function plots a histogram of the age distribution for members of a specified party.
}
\examples{
# Example usage:
# Assuming member_basic_info data is loaded
# input <- list(party_name = "Test_Party")
# output <- list(main_plot = renderPlot(function() NULL))
# draw_member_age_hist(input, output)
}
