% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{get_member_detail_info}
\alias{get_member_detail_info}
\title{Extract detailed information of members}
\usage{
get_member_detail_info(detail_url, party_member_info_df)
}
\arguments{
\item{detail_url}{A vector of member detail URLs}

\item{party_member_info_df}{The dataframe to store the extracted information}
}
\value{
An updated dataframe with member details
}
\description{
This function extracts detailed information of each member from their respective detail URLs and appends it to a dataframe.
}
