% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{get_name_return_detail_url_by_alph}
\alias{get_name_return_detail_url_by_alph}
\title{Get Member Detail URLs and Names by Alphabet}
\usage{
get_name_return_detail_url_by_alph(start_letter, html)

get_name_return_detail_url_by_alph(start_letter, html)
}
\arguments{
\item{start_letter}{The starting letter of the members' last names}

\item{html}{The HTML document containing the member list}
}
\value{
A list of member detail URLs and their corresponding full names.

A list of member detail URLs
}
\description{
Extracts member detail URLs and names from an HTML document for a given starting letter.

This function extracts the URLs and names of members from an HTML document based on the starting letter of their last name.
}
\examples{
# Example usage:
# Assume `html` is an HTML document object
html <- read_html("https://example.com/members")
get_name_return_detail_url_by_alph("A", html)

}
