% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_basic_info.R
\docType{data}
\name{member_basic_info}
\alias{member_basic_info}
\title{Member Basic Information}
\format{
A data frame with columns:
\describe{
  \item{Name}{Name of the member (character)}
  \item{Party}{Political party of the member (character)}
  \item{Constituency}{Constituency of the member (character)}
  \item{Birthyear}{Year of birth of the member (numeric)}
  \item{Address}{Address of the member (character)}
  \item{Email}{Email of the member (character)}
  \item{Telephone}{Telephone number of the member (character)}
  \item{Title}{Title of the member (character)}
  \item{Age}{Age of the member (numeric)}
}
}
\source{
\url{https://www.riksdagen.se/en/members-and-parties/members/}
}
\usage{
member_basic_info
}
\description{
A dataset containing basic information about members, including their name, party, constituency, birth year, and other details.
}
\keyword{datasets}
