% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parliments_source.R
\docType{data}
\name{parliments_source}
\alias{parliments_source}
\title{Parliaments Source URL}
\format{
A character string representing the URL to fetch parliament member information.
}
\source{
\url{https://www.riksdagen.se/en/members-and-parties/members/}
}
\usage{
parliments_source
}
\description{
The URL from which member information is fetched. This URL is used to scrape the members' details.
}
\keyword{datasets}
