% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{register_data}
\alias{register_data}
\title{Register member basic information}
\usage{
register_data(member_basic_info, party)
}
\arguments{
\item{member_basic_info}{The dataframe containing member basic information}

\item{party}{The political party of the members}
}
\description{
This function writes the member basic information to a CSV file and registers it in the project.
}
